Aplicando INSERT, UPDATE E DELETE com banco de dados

+++ src/BacanaBurguesCrud/BacanasBurgues.Entidades/Produto.cs
@@ -8,8 +8,8 @@ namespace BacanasBurgues.Entidades
         public string Nome { get; set; }
         public string Tipo { get; set; }
         public int Quantidade { get; set; }
-        public double Preco { get; set; }
-        public double Lucro { get; set; }
+        public decimal Preco { get; set; }
+        public int Lucro { get; set; }
 
         //Nome: Nome do produto
         //Tipo: Em qual tipo esse produto se enquadra
@@ -17,7 +17,7 @@ namespace BacanasBurgues.Entidades
         // Preco: O preço do produto em si, sem o o adicional.
         // Lucro : E quantos porcentos  do valor o cliente quer adicionar ao produto.
 
-        public Produto(double preco, string nome, string tipo, int quantidade)
+        public Produto(decimal preco, string nome, string tipo, int quantidade)
         {
             Preco = preco;
             Nome = nome;
